[#261] feat: 로그인 여부에 따라 에러해결
