+++ src/components/Chatting/SentHi/Teams/Styles.ts
@@ -35,7 +35,8 @@ export const NoTeamsMessageContainer = styled.div`
 `;
 
 export const ZimeetLogo = styled(I.ZimeetLogo)`
-  width: 30%;
+  width: 40%;
+  height: 16%;
   margin-bottom: 5%;
  `;
