+++ src/components/MyPage/Item/Item.tsx
@@ -5,7 +5,7 @@ const Item = () => {
   return (
     <S.IteamLayout>
         <S.BannerTitle>
-            <b>ZI밋+ 등급 | 늘품제 시즌권</b><br/>
+            <b>ZI밋+ 등급 | 아우름제 시즌권</b><br/>
             모든 팀의 상세 정보를 3주 동안 무제한 확인 가능
         </S.BannerTitle>
         <S.BannerPriceWrapper>
